Warnings in generated VhdlParser.cc
When compiling the generated VhdlParser.cc we get a number of warnings about the return type, these are no problems as they all follow an assert(false) but when "promoting" the missing to return type warning to an error (openSUSE) this is a problem.

The best way would be that the problem is solved in javaCC though this will be probably a long process, so we use the workaround to overcome the problem.
(The problem was detected as part of the reproducibility tests for openSUSE, see #8158)