Doxygen crashes on Fortran
When having a Fortran file like:
```
!>  module docu
MODULE test_mod
  INTERFACE
    !> @brief iets
    SUBROUTINE subr_i(this)
      INTEGER this
    END SUBROUTINE subr_i
  END INTERFACE

  !< @brief type brief
  TYPE, PUBLIC :: test_type
    !> docu
    integer i
  END TYPE test_type
END MODULE test_mod
```
this is due to the fact that a incorrect start of comment `!<` is used for the `TYPE` and that initiated because the last `SUBROUTINE` argument does not have any documentation.
The actual cause is that at the end of a subroutine the `vtype` is not properly reset.