XHML: div tag not possible as part of a p tag.
With test 76 we get a number of messages like:
```
element p: validity error : Element div is not declared in p list of possible children
```

this is a regression on issue #8169 / pull request #8170.
The `<div>` tag is not allowed as child of a `<p>` tag and furthermore in the output when a number of items should be on one line this is not the case anymore.
This all can be corrected by means of using the `<span>` tag instead of the `<div>` tag.