Present all warnings of first documentation sub-task when it fails
This is a follow-up on pull request #8199 (Have build of doxygen documentation fail on warnings.) and #8197 (issue #8023 Add option to exit with non-zero exit code on warnings even if WARN_AS_ERROR is OFF).
These were 2 independent situation, but as with #8197 incor[porated into doxygen it is no possible to extend #8199 to fail but at present all warnings of the first (sub-)task that fails.