Warning building test 45 i.e. incorrect links specified
When running test 45 for HTML and LaTeX  and doing link checks / build pdf we get the warnings:
```
List of broken links and other issues:
file:///.../html/item2.html
 Lines: 71, 76
  Code: 200 (no message)
 To do: Some of the links to this resource point to broken URI fragments
        (such as index.html#fragment).
The following fragments need to be fixed:
        item2                           Line: 71

```
and
```
LaTeX Warning: Reference `item2_item2' on page 1 undefined on input line 5.
```

In case a refitem references to a page it should add an anchor behind it.