issue #8246 The prefix &#х202А; for files is displayed incorrectly in the file names.
Each index generator should handle its Left-To-Right / Right-To-Left handling on its own as otherwise we get for e.g.
treeview in the js file:
```
[ "\u202A&#x202A;afx3.h", "a00002_source.html", null ]
```
instead of:
```
[ "\u202A;afx3.h", "a00002_source.html", null ]
```
resulting in the problems from this issue.

in the htmlhelp (index.hhc):
```
... value="&#x202A;&#x202A;afx3.h" ...
```
i.e. a double `&#x202A;` which is not nice.

In "XML" type formats (not the doxygen xml output!), here qhp index.qhp:
```
... title="&amp;#x202A;afx3.h" ...
```
where the LTR sequence is escaped.