Doxygen's own documentation doesn't build with recent LaTeX version
Due to a recent change in the new distribution of LaTeX in the handling of the "input" commands we get the error like:
```
Appendix D.
(../html/examples/group/latex/refman_doc.tex (../html/examples/group/latex//gro
up__group1.tex) [243]
! I can't write on file `../html/examples/group/latex//group__group2.aux'.
\@include ...mmediate \openout \@partaux "#1.aux"
                                                  \immediate \write \@partau...
l.3 \include{group__group2}

Please type another output file name
! Emergency stop.
\@include ...mmediate \openout \@partaux "#1.aux"
                                                  \immediate \write \@partau...
l.3 \include{group__group2}

*** (job aborted, file error in nonstop mode)
```

The problem is that now an intermediate file is written to `../html/examples/group/latex//group__group2.aux` that was written in the older distributions as `./group__group2.aux`, so in the current directory.
Writing to sub directories of the current directory is possible but it is not allowed (unless special, dangerous, options are used) to write to directories outside tgese directories.
(see also: https://tex.stackexchange.com/questions/575120/problem-writing-aux-file)

This fix writes the examples in subdirectories in the latex directory and not under the, parallel, html directory. This means also that there is a cleaner and clearer distinction between the html and latex output.