Layout in doxywizard documentation
When looking at page 100 of the doxygen 1.9.0 manual (i.e. chapter Doxywizard usage, bold part "Configure doxygen using the Wizard and/or Expert tab..Configure doxygen using the Wizard and/or Expert tab...") we see that the word "Wizard tab" in comparison to "Expert tab" are a bit on a strange place. In the HTML documentation this looks OK
The problem is caused by the fact that we have her nested description (`<dl>` lists), a new paragraph should be started / the line should be properly filled and this can be done by means of the `\hfill`.