Incorrect entries on treeview
When having e.g. markdown files like:
```
0.3.1.md
0.3.2.md
```
we get in the treeview twice the item "0" instead of the, more, expected "0.3.1" and "0.3.2".
This is due to the fact that the filename is seen as `0` and the extension `.3.1.md` and `.3.2.md` instead of
filename `0.3.1` and `0.3.2` and extension in both cases `.md`.

The problem was found by Fossies in the Buildbot project.