Use language name to get code coloring
In a `\code` or `~~~` or ` ``` ` environment we can define the language of a code block by specifying an extension.
In markdown we see quite often the language name as code block name (especially for `python`)
One can specify it by means of an `EXTENSION_MAPPING` but this is quite unnatural as it it not an extension.

- see to it that the extension (and thus `EXTENSION_MAPPING`) still has precedence
- in case not a known extension, try the language names.

Note: with docbook the explicit name usage was missing.