Add possibility to preserve relevant comments when using -x option
In case of using the `-x` with doxygen we get a short list with the differences compared to the default doxygen settings.

Sometimes it is useful to see also the comments of the non default items and save the user comments (`##`) of the Doxyfile.
To be able to do this the "new"  option `-xl` is introduced.
(explicitly not introducing a new single letter option as `-l` already exists and also a new option would complicate the option handling).