Wrong place in comment scanner of end summary tag
When having:
```
/// \file

/// The namespace docu
namespace Demo
{
    /// <summary>
    /// The base class
    /// </summary>
    public class Base
    {
        /// <summary>
        /// The foo function
        /// </summary>
        /// <remarks>
        /// <param name="value">The foo parameter</param>
        /// <returns>foo Something</returns>
        /// </remarks>
        public virtual int foo(int value);
    }
}
```
and we run this with `doxygen -d commentscan` we see that the `</summary>` is in the detailed part and not in the brief part:
```
CommentScanner: D:/speeltuin/issue_7356/bug_csharp_summary/Class1.cs:6
input=[
<summary>
The base class
</summary>
 ]
-----------
CommentScanner: D:/speeltuin/issue_7356/bug_csharp_summary/Class1.cs:6
output=[
brief=[line=6
<summary>
The base class
]
docs=[line=6
</summary> ]
inbody=[line=-1
]
]
```
this happened between the versions 1.8.14(OK) and 1.8.15(wrong), most likely through an improvement in case there are multiple brief section.
In case `</summary>` we should be already in the brief section so no need to switch again as the second switch would have no effect and we would land in the details.