Extension during mapping not correctly replaced.
When searching whether the extension is already in the list of mapped extensions the input extension is used but when inserting the eventually corrected extension (extName) is used.
In all cases the inserted extension contains a dot (`.`) , but for the input extension this is not necessarily the case. This can cause problems when searching for an extension as the search will always search wit a dot and thus find the first (is old) value.