Parser for verbatim including a (text) file.
In older projects it is quite common to have files like `README.txt` that don't contain code or markdown markup, but that would be nice to have verbatim into the documentation.
The same is true for other types of files like configuration files or sh / zsh / bat files that should not be processed in ant way.

For this the "language" `unparsed` is created.