Adding support for lex files
Adding support for lex. The lex files consist of a mixture of lex specific structures, that are handled in the the lexcode.l and lexscanner.l, abd C code that is handled by the C parsers (hence the rules used are partly copied from scanner.l).

Special attention has been paid to memberdef.cpp as the initial values should not be handled by the lex parsers but by the C parsers.