Strange HTML tab title in case of a page without title
In case we have an example like:
```
/// @page Bug_err
///
/// \brief Some text in cpp file
/// \details Some detailed text in cpp file
```
we will have as in the HTML tab: `Project Title: $title` as the page has an empty title.
A better title would in this case be: `Project Title: Bug_err`.

This has been implemented  by setting the name of the page in the empty title. In case of a main page the xml output sets the empty title to the PROJECT_NAME, as the new, default, setting would see to it that also the mainpage with an empty title would be set we have to see that this is done beforehand.