Warnings regardin9 Internal inconsistency: namespace in IDL
In the ACE TAO package we find a number of warnings like:
```
.../ACE_TAO/ACE/html/libtao-doc/security/TAO_Security.tag:1: error: Internal inconsistency: namespace in IDL not module, library or constant group
.../ACE_TAO/TAO/orbsvcs/orbsvcs/CosEventChannelAdmin.idl:30: error: Internal inconsistency: namespace in IDL not module, library or constant group
```

There are a number of reasons for these warnings a.o. IDL modules not handled properly but seen as namespaces and when type is not an IDL modules a warning is given. This happens especially when the module is documented with `@namespace`, this can partly be solved by testing the file type but when in a general file (out of place documentation) this does not work.
- introducing new commands to document IDL specific "namespaces"
- introducing and handling of IDL specific "namespaces" in tag files.

(A problem remains when we have e.g. an IDL module and in a C++ file there is a namespace with the same name.)