Warnings regarding Internal inconsistency: namespace in IDL
n the ACE TAO package we find a number of warnings like:
```
.../ACE_TAO/ACE/html/libtao-doc/security/TAO_Security.tag:1: error: Internal inconsistency: namespace in IDL not module, library or constant group
.../ACE_TAO/TAO/orbsvcs/orbsvcs/CosEventChannelAdmin.idl:30: error: Internal inconsistency: namespace in IDL not module, library or constant group
```

There are a number of reasons for these warnings a.o. IDL modules not handled properly but seen as namespaces and when type is not an IDL modules a warning is given.
This happens especially when the module is documented with `@namespace` or there is a namespace definition with the same name in a C++ file (which is just a mapping between IDL and C++).

Based on the discussion in https://github.com/DOCGroup/ACE_TAO/pull/1450#issuecomment-799277574 and the results from pull request #8429
a new version has been crated with just the extra test this just removes all Internal inconsistency warnings.