Incorrect handling of JAVA style code statement in preprocessor
When we have a program like:
```
/**
 * {@code
 * uspoof_setChecks(USPOOF_ALL_CHECKS & ~USPOOF_CONFUSABLE);
 * }
 *
 */
U_CAPI void U_EXPORT2
uspoof_setChecks(USpoofChecker *sc, int32_t checks, UErrorCode *status);

/**
 * \cond
 */
U_DEFINE_LOCAL_OPEN_POINTER(LocalUSpoofCheckerPointer, USpoofChecker, uspoof_close);
/** \endcond */
```
In the console output this results in some empty lines from the commentcnv lexer.

The problem comes from the preprocessor that doesn't recognize `{@code` as a JAVA style code statement that should end at the corresponding `}` and thus we remain in a code section. and thus not handling the `\cond` statement.
In the commentcnv lexer the `{@code` is handled as a special case and this is done in the preprocessor now as well.