Automatically create multi-layer subdirectories
When have a Doxyfile like:
```
QUIET=YES
OUTPUT_DIRECTORY = tst/lev1
HTML_OUTPUT = html_1/h_lev1
LATEX_OUTPUT = latex_1/l_lev1/l_lev2
```
we get the warning:
```
error: tag OUTPUT_DIRECTORY: Output directory 'tst/lev1' does not exist and cannot be created
```

In the previously used QDir  software there was only a `mkdir` routine and not a `mkpath` routine (as in newer Qt versions).
With the new implementation of the QDir through the ghc filesystem it is, easily, possible to create a `mkpath` by means of a `create_subdirectories`

Note: it is not necessary to replace all `mkdir` by `mkpath` as at some places it is guaranteed that only one level of directories can be missing (but it won't hurt).