Write out used settings when generating XML output
When generating XML output is is afterwards unknown which settings have been used to generate the XML output, e.g. whether `EXTRACT_PRIVATE` was set or not as the XML output tries to write out all information contained in the sources and not to limit it.
By writing out a Doxyfile.xml with all the used settings this can be overcome.