Inconsistency between files with underscore or in sub directories
When having the files:
```
docs_dev_git.md
docs/dev/git.md
```
with at the beginning something like:
```
# The header
```
we get the waring:
```
docs_dev_git.md:1: warning: multiple use of page label 'md_docs_dev_git', (other occurrence: <md_docs_dev_git>, line: 1)
```

This due to the fact that the "/" is translated to a "_" and thus clashes.
This is fixed by means of converting an existing "_" in the filename etc. to "__"