issue #8518 tag </programlisting> was inserted before </highlight> parsing python file to xml
Doxygen expects that decorators don't have a return in them, but the complicated decorator:
```
   @basetest.unittest.skipIf(
      api_implementation.Type() != 'cpp' or api_implementation.Version() != 2,
       'Errors are only available from the most recent C++ implementation.')
```
has some in them.
This leads to the mentioned problem as well as that the line counting is incorrect (we see that the last mentioned lie number in the xml file is 2964 though the file has 2966 lines). By fixing this (and also the line counting in the python scanner) the problem looks to be handled.