Confusing debug output for markdown
When having debug output one expects to have an overview what goes into the markdown processor and comes out of it. For markdown this happens but there can be a small step in front of it (page handling) and than the input is confusing as some processing already took place.

When having a file aa.md:
```
This is a .md file
```
we get with `doxygen -d markdown`:
```
======== Markdown =========
---- input -------
@page md_aa aa\ilinebr This is a .md file

---- output -----
@page md_aa aa\ilinebr This is a .md file

=========
```
whilst it would be less confusing when we have:
```
======== Markdown =========
---- input -------
This is a .md file

---- output -----
@page md_aa aa\ilinebr This is a .md file

=========
```