No specific way to set class attributes of a treeview link.
A link in a navtree / treeview has not a general class attached to it so its properties can only be changed by means of the global `a` tag.
There is a class attached to it but this contains the link reference so it is possible to highlight an item in the treeview when clicking in the text part of the documentation on a link.
A general class has been added to the navtree link.