issue #8564 input buffer overflow, can't enlarge buffer because scanner uses REJECT
When we have a very large "LONGSTRINBLOCK" (e.g. a very large documentation part or an encode executable in e.g. base64 encoding) , there are a lot of lines that have to be read at once, this is not really necessary they can be read sequentially (probably at the cost of some speed)