#8569 Java, xml: members after certain comments with {@code} are missing
The handling of special commands in normal comment blocks (contrary to doxygen comment block starting wit `///`, `//!`, `/**` and `/*!`) was done for compatibility with, the now obsolete / unmaintained / dead package Doc++.
This compatibility was already dropped for some commands / languages but will now be dropped in general.

This patch handles the situation wit non-doxygen comments `//`, the situation with `/*` will be handled separately.