Running tests with subdirs enabled
When running the standard doxygen tests with subdirs enabled (i.e. on windows `nmake tests TESTS_FLAGS==--subdirs`) we get the following type errors:
```
Failed to run xmllint on the doxygen output file ..../testing/test_output_008
```
this is due to the fact that the file to be checked is in a subdirectory and as directory separator a backslash is used (initially there are 2 backslashes but due to e.g. `check_file = check_file[0]` one backslash disappears. Furthermore everywhere else the backslashes are converted to forward slashes.

In the resulting files there are references to the mentioned sub directories path, but the file to compare to is without therse so we have to filter them as well.