Problem with refitem command (tests 45)
When running a link checker over the (x)html results of the doxygen tests 45 we get the error:
```
===============================

Processing      file:///.../ex_1/html/index.xhtml

List of broken links and other issues:
file:///.../ex_1/html/item2.xhtml
  Line: 71
  Code: 200 (no message)
 To do: Some of the links to this resource point to broken URI fragments
        (such as index.html#fragment).
The following fragments need to be fixed:
        item2                           Line: 71

```

This accounts also for the second example (just with subdirs enabled).

When we change *.dox files so that the `\mainpage` is disabled and `\page` is enabled the result for the subdir version is so that the links don't work at all as it searches for e.g. `.../html/d5/d4a/d5/d4a/pg1.xhtml#item1` instead of `html/d5/d4a/pg1.xhtml#item1`.

(I slightly modified the example as to compare the links with the links for the `\ref` command).