Implementation functions for INLINE_SIMPLE_STRUCTS for docbook output
The implementation of the `INLINE_SIMPLE_STRUCTS` functions was missing for docbook.
The problem shows up when using e.g.
```
/** \file */

/** outer */
struct Outer
{
  /** foo */
  union Foo
  {
    /** Bar */
    struct FooFlags
    {
      bool cond1; /*!< \brief bar 1
                   *   \details details1 bar 1
                   */
      bool cond2; /*!< bar 2 */
    } flags; /*!< \brief foo bar
              *   \details details2 of foo bar
              */
  } myMember; /*!< public member */
private:
  void myWork(); /*!< private member function */
};
```
with
```
INLINE_SIMPLE_STRUCTS  = YES
QUIET=YES
GENERATE_DOCBOOK=YES
```

Note with the `INLINE_SIMPLE_STRUCTS` there is also a mismatch with the opening  and closing `section` tags, but that is unrelated to this implementation(in this case is is "solved" by adding a closing section tag to struct_outer.xml and removing the last closing section tag from structs_8hpp.xml