Missing anchors to source code in RTF
When running the doxygen test 25, like `make tests TEST_FLAGS="--rtf  --id=25 --keep --noredir --cfg RTF_HYPERLINKS=YES"`) and clicking on the `example_test.cpp` a jump occurs to the top of the document and not to the (available) source code.
- corrected generation of anchors (not the name was important for the anchor but the fileName.
- adding test possibility for detecting missing anchors in test procedure

We should also check whether or not a bookmark is double defined, but test 57 (inline namespace) gives a, small, problem here (in all formats!). This has to be handled in a separate issue.