Incorrect docbook output (section tag  mismatch)
In case `INLINE_SIMPLE_STRUCTS` has been set the resulting opening / closing  section tags don't match.
This problem can be be seen with the doxygen tests 36, 67 and 68 and well as with the test case used for issue #8588
Instead of explicit using special flags now a general close is done (in the cases examined no problems were found with this strategy).