bug_744840  <tt> blocks cause error "end of comment while expecting command </code>"
The problem is cased due to setting of auto brief options, and the fact that `...` is not seen as a 'special character' but as the end of a brief comment.
(In the current master the waring is a bit better already:
```
warning: end of comment block while expecting command </tt>
```
but the root cause is still present).

Making the `....` to be seen as a 'special character'