Incorrect line number in warnings in case of if command
When having a problem like:
```
/// \file

/**
 * aaa
 */
typedef enum {
  /** bbb
   * \line_9
   * @if INTERNAL
   * \line_11
   * @endif
   * \line_13
   */
         BS
} bsb_type_enum;
```
with Doxyfile:
```
JAVADOC_AUTOBRIEF      = NO
ENABLED_SECTIONS       = INTERNAL
GENERATE_LATEX         = NO
INPUT=b.h
QUIET                  = YES
```
we get the warnings:
.../b.h:9: warning: Found unknown command '\line_9'
.../b.h:10: warning: Found unknown command '\line_11'
.../b.h:11: warning: Found unknown command '\line_13'
```
so the line number in the warning is incorrect for the lines after the `if` and `endif` command.
Looking at the output of the commentscanner we see:
```
CommentScanner: .../b.h:8
output=[
brief=[line=-1
]
docs=[line=8
bbb
 \line_9
 \line_11
  \line_13]
inbody=[line=-1
]
]
```
in other words a number of lines "disappear".

As the `if` command is an inline command it is not possible to just replace e.g. the `if` line wit an empty line (as this would be seen as a paragraph break).

With this patch a new internal command `iline` is defined with which it is possible to overrule in the documentation tokenizer the line counter.