issue #8642 invalid argument for command iline
From the string presented to the tokenizer the end spaces are stripped so it is not necessary that a soace is present at the end.
Due to the brief command it is also possible tat a `.` is present at the end.
These cases are handled now.

Furthermore small improvement on warning message.