Warning message in case of preprocesssor macro substitution
In the macro preprocessor it is not allowed to have `##` at the beginning of a definition nor at the end.
Furthermore a single `#` should be followed by an ID (now enforced by means of the ID-start  character).

Normally in case we have something like (derived from berkeley_upc-2021.4.0 package):
```
#define GASNET_ALIGNED_SEGMENTS   ###

#if !defined(GASNET_ALIGNED_SEGMENTS) || \
    (defined(GASNET_ALIGNED_SEGMENTS) && GASNET_ALIGNED_SEGMENTS != 0 && GASNET_ALIGNED_SEGMENTS != 1)
  #error GASNet core failed to define GASNET_ALIGNED_SEGMENTS to 0 or 1
#endif
```
we get as warning:
```
...:4: warning: preprocessing issue while doing constant expression evaluation: syntax error: input=' ! 1L  ||      ( 1L  && #  != 0 && #  != 1)'
```
but nothing about the real problem (i.e. incorrect define) and now we also get:
```
...:1: warning: '##' cannot occur at the beginning of a macro definition 'GASNET_ALIGNED_SEGMENTS': '###'
```
so it is a bit easier to find the problem.