Handling escaped own end block command in special command block
When having an example like:
```
# doxygen

@htmlonly
* finding \\endhtmlonly when
* \code
@endhtmlonly

@code
* finding \\endcode when
* \verbatim
@endcode
```

we get warnings like:
```
.../aa.md:13: warning: reached end of comment while inside a \verbatim block; check for missing \endverbatim tag!
.../aa.md:12: warning: verbatim section ended without end marker
```
this due to the fact that inside the special blocks the `\` is eaten separately and afterwards the parser sees the own special block end command and interpret this.
By handling the escaped command character (so `\\` etc.) separately we can overcome this problem.
(a workaround might be that the user uses e.g. `\&zwj;\`)