Add config option SHOW_HEADERFILE
This adds a new configuration option that controls whether to show the
file specified by \headerfile (or the header-file argument of \class) in
the documentation for a class or concept.

Previously that was controlled by SHOW_INCLUDE_FILES, but that should be
used for showing the include list for \file documentation. With this
change, there are two separate options to control the two separate
behaviours.

Fixes #8639