issue #7434 Error in LaTeX output of doxygen documentation of VHDL
In this issue there are 2 texts to be replaced: Class and Namespace. Most Class texts are already corrected, though, at least, the sentence "The documentation for this class was generated from the following file:" was not corrected.
Based on the output languages where the word "design unit" was translated the relevant function was, best guess, corrected.