Warnings running doxygen on its own internal documentation
When we run doxygen on its own sources we te errors:
```
error: Found invalid bounding FIG -1 in file .../doxygen_docs/html/d0/d32/class_dot_graph.html!
error: Found invalid SVG id in file .../doxygen_docs/html/db/daa/class_dot_file_patcher.html!
error: Found invalid bounding FIG -1 in file .../doxygen_docs/html/db/daa/class_dot_file_patcher.html!
```
tis is due to the fact that the INLINE_SOURCES`  is set and we have the doxygen sources in the HTML files. For the dot file patcher in these files there is a search for strings to be patched with the correct image commands (files that are generated by dot).
Due to the simplicity of the test also the search / replace text was found with the test., the test has been extended a bit so that a better search is performed.