bug_624366 Expanded seealso tags create multiple "See also" sections
Also known as issue #3850

Although in the original specs of the documentation tags it would be better when both both types of closing tags (`<.../>` and `<...></..>`) are supported.
In newer versions of the Microsoft specifications (see e.g. https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/xmldoc/recommended-tags#seealso) also a text is supported in the seealso tag.