Warning with python file starting and ending with __ as file name
When we have a file in python with a file name starting and ending with `__` other than `__init__,.py` we get a message like:
```
warning: found </strong> tag without matching <strong>
```
this is due to the fact that the file name is translated to a namespace with name like the file name and this is translated  to e.g. `<strong>future</strong>`.