access attribute resolution
+ Resolves the module "public/private" object recognition issue by simply removing
all cases of private access recognition in Doxygen. Indeed, it is not really clear why such
private access recognition has been added to Doxygen in the first place as it seems entirely
unnecessary. The problem is that any public/private statement with followup object names,
messes with the objects' hyperlinking in the code, unless each object separately has its own
access attribute declared. Unless the attribute is public, the object hyperlinks to documentation
pages are disabled, which seems unnecessary.

+ Also resolves the `bind(C)` attribute highlighting issue.