Incorrect warning regarding SQLITE3 seyyttings
When we have in the settings files settings that are not enabled in the current doxygen version we get for the CLANG part
```
warning: Tag 'CLANG_DATABASE_PATH' at line 391 of file 'Doxyfile' belongs to an option that was not enabled at compile time.
         To avoid this warning please remove this line from your configuration file or upgrade it using "doxygen -u", or recompile doxygen with this feature enabled.
```
though for the SQLITE3 part we get:
```
warning: ignoring unsupported tag 'GENERATE_SQLITE3' at line 392, file Doxyfile
```
this is due to the fact that the info of the group was closed after the complete group and not directly after the info.