value attribute for <li> tag
With the `<li>` tag in HTML it is possible to have a `value=` attribute, this was only working for the HTML output, but has now been added to other output formats as well.
Also fixed small counting issue in case we have more than the supported number of levels.

(For docbook there is no solution yet)