value attribute for <li> tag for docbook
In the docbook listitem tag there is no option for the implementation of the value attribute.
We have to use a workaround as suggested from the support person of XMLMind: close the current ordered list and start a new list (i.e. at the same level).
As we need to know the type of list (Arabic numbers, lower / upper case Roman) we have to "reparse" the attributes of the List again.
Also special care has to be taken when we have a value attribute with the first list item (we have to prevent that we get an empty ordered docbook list (this results in the m_isFirst field with the list items).