bug_684050 line continuation in \copydoc
Conform the documentation of `\copydoc` it cannot have an embedded linefeed and thus the `\copydoc` is terminated here when converted to \copybrief` and `copydetails`.
The rest of the "signature) is automatically added at the `\copydetails` leaving the `copybrief` with just half a line (all can be observed with `doxygen -d commentscan`).
The round brackets were not handled and now a rudimentary  version has been implemented.

(Note the original issue contains 3 problems this is the last one of the 3 and about copydoc)