Name of used struct
In the doxygen the struct name `CondCtx` is used twice (pre.l and commentcnv.l), although syntactically correct and no intersection between the 2 structs (and they are vene identical) the internal doxygen  documentation has some problems (especially we get multiple times the same label defined).
Analogous to the `state` struct it is better to use an unique name.