Correcting labels
In case we have included sources (e.g. due to `INLINE_SOURCES`) and also the `SOURCE_BROWSER` enabled it is possible that we get double labels.
Also the calculation of labels in RTF was not done consistently, sometimes the path was striper and on other moments it wasn;t (especially seen in case CREATE_SUBDIRS
The problem was found using RTF / Docbook on the doxygen internal documentation (not standard) but will happen too in other projects