Correction of alphabetical lists in LaTeX output
There was a small discrepancy when using the default list in LaTeX and when specifying the type of list.
In the  default case still the `\alph*` and `\Alph*` representations were used resulting in problems when the number of items exceeded the6. In case we use a list with `type=`a"` or `type="A"` this problem was solved by using the dedicated lists `\enumaphaphcnt*` and `\enumAlphAlphcnt*`,.
Now always the dedicated lists are used.