Don't use full path for *file in XML output
When having:
```
\dotfile overview.dot
```
we se in the XML output e.g.
```
<para><dotfile name="<full_path>/overview.dot"></dotfile>
```
though it is not guaranteed that this is a persistent path.
Copying the file to the xml directory makes it local to the xml directory and this next to the generated xml output and the reference can be:
```
<para><dotfile name="overview.dot"></dotfile>
```
Similar will be the case for `\diafile` and \mscfile`