Copy original dotfile to output directory in case of DOT_CLEANUP is not set
In case of a `\dot` image the dot file is created and when `DOT_CLEANUP` is not set the file is left in the output directory.
In case of the `\dotfile` and the `DOT_CLEANUP` is not set the file is not placed in the output directory as the image can be generated from the original file.
This is a bit inconsistent as we would have for some images a `dot` file and for others we wouldn't.

Similar reasoning for msc and dia