issue #8854 Class hierarchy diagram is not correct, if inherited from Python built-in class ABC
The previous patch by luck solved the problem.
The actual problem is that only one subgraph was run and this did work when the root node was known to doxygen but here this is not really the case (it is from an external source).
The renumbering was introduced in (fb66cae35cd59d1026fe3c6de46ae3a2ed6f9f91 ):
```
Commit: fb66cae35cd59d1026fe3c6de46ae3a2ed6f9f91 [fb66cae]
Commit Date: Tuesday, February 2, 2016 9:13:44 PM
Reimplemented node renumbering for dot graph to improve performance
```
and the problem was present from this point on.