Warning on empty \cond statement plus extra empty line.
When we have a Fortran file like:
```
!> \file

!> \cond
!>
      subroutine test
      end subroutine
!> \endcond
```
We get the warning:
```
.../bb.f90:4: warning: problem evaluating expression '!': Unexpected end of expression
```
But in the console output we also see an empty line like:`
```
Reading .../bb.f90...

.../bb.f90:4: warning: problem evaluating expression '!': Unexpected end of expression
Prepassing fixed form of .../bb.f90
```
This newline can also be observed in C code, when the preprocessing is disabled.
This is due to the fact that during the processing of the `CondLine` the `\n` is not caught properly and thus the default rule kicks in.