Incorrect handling of JAVA style code statement in preprocessor
When we have a program like:
```
/**
 * {@code
 * fie
 * }
 */
void fie();

/**
 * \cond
 */
U_DEFINE_LOCAL_OPEN_POINTER(LocalUSpoofCheckerPointer, USpoofChecker, uspoof_close);
/** \endcond */
```
and run `doxygen -d preprocessor` we see that the `\cond` part is not removed.
The problem comes from the preprocessor that doesn't recognize `{@code` as a JAVA style code statement that should end at the corresponding `}` and thus we remain in a code section. and thus not handling the `\cond` statement.
In the commentcnv lexer the `{@code` is handled as a special case and this is done in the preprocessor now as well.

The `{@code` has also been brought in line with the other JAVA style commands by allowing white space between `{` and `@`.